backend/requirements.txt
django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.8.2
requests==2.32.3
